test: unskip error-location test and add shared conftest fixtures
- Remove the unconditional `@pytest.mark.skip` on
  `test_validation_error_location`; the assertion only checks that "name"
  appears in the reported `loc`, which is tolerant of Pydantic v2 error-format
  differences, so the path is now exercised on all supported versions instead
  of skipped everywhere.
- Add `tests/conftest.py` housing the shared `mock_request_factory` fixture
  (moved from `test_pipeline.py`) so it is auto-discovered by every test
  module, matching the toolkit convention (langgraph/durable-graph/db).

No production code changes. Full suite passes with coverage at 96%.

Closes #219
